refactor(policy): merge core/policy_engine into policy/
