--- src/ml_lifecycle_platform/policy/__init__.py
@@ -1,1 +1,0 @@
-"""Release policy evaluation. Logic lives in core.policy_engine."""
